requirements_test.txt
pytest-homeassistant-custom-component>=0.13.102
pylint
mypy
types-requests
types-setuptools
protobuf
pytest-cov
